persist apple auth
