+++ frontend/app/(authenticated)/(tabs)/_layout.tsx
