+++ frontend/app/(authenticated)/(tabs)/explore.tsx
