+++ frontend/app/(authenticated)/(tabs)/index.tsx
